[Refactor] : ADModule to plugin
Git : #95
