+++ .mise.toml
@@ -1,2 +1,2 @@
 [tools]
-tuist = "4.10.1"
+tuist = "4.20.0"
