--- .tuist-version
@@ -1,1 +1,0 @@
-4.10.1
