--- Plugins/ProjectEnvironment/Plugin.swift
@@ -1,3 +1,0 @@
-import ProjectDescription
-
-let plugin = Plugin(name: "ProjectEnvironment")
