+++ Tuist/Plugins/ADModules/Layer/Domain/ImageTools.swift
