+++ Tuist/Plugins/ADModules/Layer/Presentation/CropImage.swift
