+++ Tuist/Plugins/ADModules/Layer/Presentation/MaskImage.swift
