+++ Tuist/Plugins/ADModules/Plugin.swift
@@ -1,0 +1,10 @@
+//
+//  Plugin.swift
+//  ADAsyncAlgorithmsManifests
+//
+//  Created by chminii on 7/3/24.
+//
+
+import ProjectDescription
+
+let plugin = Plugin(name: "ADModules")
