+++ Tuist/ProjectDescriptionHelpers/Configuration.swift
