+++ Tuist/ProjectDescriptionHelpers/InfoPlist.swift
