+++ Tuist/ProjectDescriptionHelpers/String.swift
